finished the project
finished adding all the style properties that were needed to achieve
the desired result to finish this task. I'm sure the css file could
be cleaned up better but I'm happy it works and may return to this
project to see if I can refine my skills at coding. I solved a issue
I was having with the four pictures in the middle section. it took a bit
but I understand my initial failure and learned from it.